add helper methods for views, breadcrumbs & facets
