+++ app/controllers/catalog_controller.rb
@@ -97,7 +97,7 @@ class CatalogController < ApplicationController
     config.add_facet_field 'geogname_sim', label: 'Place', limit: 10
     config.add_facet_field 'places_ssim', label: 'Places', show: false
     config.add_facet_field 'access_subjects_ssim', label: 'Subject', limit: 10
-    config.add_facet_field 'campus_unit_sim', label: 'Campus', limit: 10
+    config.add_facet_field 'campus_unit_sim', label: 'Campus', helper_method: :render_campus_facet
     # Have BL send all facet field names to Solr, which has been the default
     # previously. Simply remove these lines if you'd rather use Solr request
     # handler defaults, or have no facets.
