+++ app/helpers/arclight_helper.rb
@@ -4,6 +4,8 @@
 ##
 # Generic Helpers used in Arclight
 module ArclightHelper
+  include CampusHelper
+
   def aria_hidden_breadcrumb_separator
     safe_join(
       [
@@ -17,7 +19,7 @@ module ArclightHelper
   ##
   # @param [SolrDocument]
   def parents_to_links(document)
-    breadcrumb_links = [add_campus(document)]
+    breadcrumb_links = [add_campus_link(document)]
     breadcrumb_links << build_repository_link(document)
     breadcrumb_links << document_parents(document).map do |parent|
       link_to parent.label, solr_document_path(parent.global_id)
@@ -34,7 +36,7 @@ module ArclightHelper
   #    shown in the mockup above. The repository and the collection parts are
   #    linked as usual; the ellipses is not linked.
   def regular_compact_breadcrumbs(document)
-    breadcrumb_links = [add_campus(document)]
+    breadcrumb_links = [add_campus_link(document)]
     breadcrumb_links << build_repository_link(document)
 
     parents = document_parents(document)
@@ -323,16 +325,18 @@ module ArclightHelper
 
   private
 
-  def build_repository_link(document)
-    repository_path = document.repository_config&.slug
-    if repository_path.present?
-      link_to(document.repository, arclight_engine.repository_path(repository_path))
-    else
-      content_tag(:span, document.repository)
+    def build_repository_link(document)
+      repository_path = document.repository_config&.slug
+      if repository_path.present?
+        link_to(document.repository, arclight_engine.repository_path(repository_path))
+      else
+        content_tag(:span, document.repository)
+      end
     end
-  end
 
-  def add_campus(document)
-    content_tag(:span, document.campus)
-  end
+    # TODO - convert to link when campus route is added
+    def add_campus_link(document)
+      # TODO campus doesn't seem to be part of this document
+      content_tag(:span, convert_campus_id(document.campus))
+    end
 end
